+++ src/debug/core/null_debug_runtime.js
@@ -1,0 +1,10 @@
+class NullDebugRuntime {
+  setLiveData() {}
+  setFightData() {}
+  setNetRoll() {}
+  printModule() {}
+  requestModule() {}
+  printEnabled() {}
+}
+
+window.NullDebugRuntime = NullDebugRuntime;